refactor and improve code

+++ data/src/main/java/com/casecode/data/repository/ItemRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.casecode.data.repository
 
+import com.casecode.data.mapper.asExternalModel
 import com.casecode.data.utils.AppDispatchers
 import com.casecode.data.utils.Dispatcher
 import com.casecode.domain.model.users.Item
@@ -7,9 +8,10 @@ import com.casecode.domain.repository.AddItem
 import com.casecode.domain.repository.AuthService
 import com.casecode.domain.repository.DeleteItem
 import com.casecode.domain.repository.ItemRepository
-import com.casecode.domain.repository.Items
+import com.casecode.domain.repository.ResourceItems
 import com.casecode.domain.repository.UpdateItem
 import com.casecode.domain.utils.ITEMS_COLLECTION_PATH
+import com.casecode.domain.utils.ITEM_NAME_FIELD
 import com.casecode.domain.utils.Resource
 import com.casecode.domain.utils.USERS_COLLECTION_PATH
 import com.casecode.pos.data.R
@@ -18,6 +20,10 @@ import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.ListenSource
+import com.google.firebase.firestore.MetadataChanges
+import com.google.firebase.firestore.SetOptions
+import com.google.firebase.firestore.SnapshotListenOptions
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
@@ -37,11 +43,17 @@ import kotlin.coroutines.suspendCoroutine
  * @property ioDispatcher Coroutine dispatcher for performing operations asynchronously on IO-bound threads.
  * @constructor Creates an [ItemRepositoryImpl] with the provided [firestore] and [ioDispatcher].
  */
-class ItemRepositoryImpl @Inject constructor(
+class ItemRepositoryImpl
+@Inject constructor(
     private val auth: FirebaseAuth,
     private val firestore: FirebaseFirestore,
     @Dispatcher(AppDispatchers.IO) val ioDispatcher: CoroutineDispatcher,
 ) : ItemRepository, AuthService {
+    // ISSUE: improve minimize cost of use server with use offline cache to get document from
+    private val optionsCache by lazy {
+        SnapshotListenOptions.Builder().setMetadataChanges(MetadataChanges.INCLUDE)
+            .setSource(ListenSource.CACHE).build()
+    }
     override val currentUserId: String
         get() = auth.currentUser?.uid.orEmpty()
 
@@ -52,39 +64,32 @@ class ItemRepositoryImpl @Inject constructor(
             awaitClose { auth.removeAuthStateListener(listener) }
         }.flowOn(ioDispatcher)
 
-    override fun getItems(): Flow<Items> =
-        callbackFlow<Resource<List<Item>>> {
-            trySend(Resource.Loading)
+    override fun getItems(): Flow<ResourceItems> = callbackFlow<Resource<List<Item>>> {
+        trySend(Resource.Loading)
 
-            val itemMutableList = mutableListOf<Item>()
+        val listenerRegistration = getItemCollectionRef(currentUserId).orderBy(ITEM_NAME_FIELD)
+            .addSnapshotListener(MetadataChanges.INCLUDE) { snapshot, error ->
 
-            // Create a Firestore query to retrieve all items for the given UID
-            val query = getItemCollectionRef(currentUserId)
-
-
-            // Query Firestore to retrieve all items for the given UID
-            val listenerRegistration = query.addSnapshotListener { snapshot, error ->
                 if (error != null) {
-                    // Handle the error
-                    Timber.tag(TAG).e(error)
+                    Timber.e(error)
                     trySend(Resource.error(R.string.error_fetching_items))
-                    return@addSnapshotListener
+                    close()
                 }
 
-                // Clear the list before updating it with new data
-                itemMutableList.clear()
-
-                // Loop through the query results and convert each document to an Item object
-                for (document in snapshot?.documents ?: emptyList()) {
-                    val item = document.toObject(Item::class.java)
-                    item?.let {
-                        // Retrieve the imageUrl from the document data
-                        val imageUrl = document.getString("image_url")
-                        val unitOfMeasurement = document.getString("unit_of_measurement")
-                        // Set the retrieved (imageUrl, unitOfMeasurement) to the Item object
-                        it.imageUrl = imageUrl
-                        it.unitOfMeasurement = unitOfMeasurement
-                        itemMutableList.add(it)
+                // Trace for where data getting form local db or server or cache memory.
+                val source = if (snapshot != null && snapshot.metadata.hasPendingWrites()) {
+                    LOCAL_DB
+                } else if (snapshot != null && snapshot.metadata.isFromCache) {
+                    LOCAL_CACHE
+                } else {
+                    SERVER_DB
+                }
+                // Handle the snapshot
+                val itemMutableList = mutableListOf<Item>()
+                with(itemMutableList) {
+                    snapshot?.documents?.mapNotNull { document ->
+                        Timber.i("$source , Item:  ${document.data}")
+                        document.toObject(Item::class.java)?.let { add(it) }
                     }
                 }
 
@@ -95,43 +100,35 @@ class ItemRepositoryImpl @Inject constructor(
                 }
             }
 
-            // Close the listener when the flow is cancelled
-            awaitClose { listenerRegistration.remove() }
-        }.flowOn(ioDispatcher)
+        // Close the listener when the flow is cancelled
+        awaitClose {
+            listenerRegistration.remove()
+        }
+    }.flowOn(ioDispatcher)
+
 
     override suspend fun addItem(item: Item): AddItem {
         return withContext(ioDispatcher) {
             try {
-                // Loading state before starting the deletion operation
-                Resource.Loading
-
                 suspendCoroutine { continuation ->
-                    val itemMap =
-                        mapOf(
-                            "name" to item.name,
-                            "price" to item.price,
-                            "quantity" to item.quantity,
-                            "sku" to item.sku,
-                            "unit_of_measurement" to item.unitOfMeasurement,
-                            "image_url" to item.imageUrl,
-                        )
-
+                    val itemMap = item.asExternalModel()
+                    if (currentUserId.isBlank()) {
+                        continuation.resume(Resource.error(com.casecode.pos.domain.R.string.uid_empty))
+                        return@suspendCoroutine
+                    }
                     val sku = item.sku
-                    getItemDocumentRef(uid = currentUserId, sku = sku).set(itemMap as Map<*, *>)
+                    getItemDocumentRef(uid = currentUserId, sku = sku).set(itemMap)
                         .addOnSuccessListener {
-                            continuation.resume(Resource.Success(data = R.string.item_added_successfully))
+                            continuation.resume(Resource.Success(R.string.item_added_successfully))
                         }.addOnFailureListener { failure ->
-                            Timber.tag(TAG).e(failure)
-
-                            val errorMessage = when (failure) {
-                                is UnknownHostException -> R.string.add_item_failure_network
-                                else -> R.string.add_item_failure_generic
-                            }
-                            continuation.resume(Resource.error(errorMessage))
+                            Timber.e(failure)
+                            continuation.resume(Resource.error(R.string.add_item_failure_generic))
                         }
                 }
+            } catch (e: UnknownHostException) {
+                Resource.error(R.string.add_item_failure_network)
             } catch (e: Exception) {
-                Timber.tag(TAG).e(e)
+                Timber.e(e)
                 Resource.error(R.string.add_item_failure_generic)
             }
         }
@@ -140,36 +137,29 @@ class ItemRepositoryImpl @Inject constructor(
     override suspend fun updateItem(item: Item): UpdateItem {
         return withContext(ioDispatcher) {
             try {
-                // Loading state before starting the update operation
-                Resource.Loading
 
                 suspendCoroutine { continuation ->
-                    val itemMap =
-                        mapOf(
-                            "name" to item.name,
-                            "price" to item.price,
-                            "quantity" to item.quantity,
-                            "sku" to item.sku,
-                            "unit_of_measurement" to item.unitOfMeasurement,
-                            "image_url" to item.imageUrl,
-                        )
-
+                    if (currentUserId.isBlank()) {
+                        continuation.resume(Resource.error(com.casecode.pos.domain.R.string.uid_empty))
+                        return@suspendCoroutine
+                    }
+                    val itemMap = item.asExternalModel()
                     val sku = item.sku
-                    getItemDocumentRef(uid = currentUserId, sku = sku).set(itemMap as Map<*, *>)
-                        .addOnSuccessListener {
-                            continuation.resume(Resource.Success(data = R.string.item_updated_successfully))
-                        }.addOnFailureListener { failure ->
-                            Timber.tag(TAG).e(failure)
+                    getItemDocumentRef(uid = currentUserId, sku = sku).set(
+                        itemMap,
+                        SetOptions.merge(),
+                    ).addOnSuccessListener {
+                        continuation.resume(Resource.Success(R.string.item_updated_successfully))
+                    }.addOnFailureListener { failure ->
+                        Timber.e(failure)
 
-                            val errorMessage = when (failure) {
-                                is UnknownHostException -> R.string.update_item_failure_network
-                                else -> R.string.update_item_failure_generic
-                            }
-                            continuation.resume(Resource.error(errorMessage))
-                        }
+                        continuation.resume(Resource.error(R.string.update_item_failure_generic))
+                    }
                 }
+            } catch (e: UnknownHostException) {
+                Resource.error(R.string.update_item_failure_network)
             } catch (e: Exception) {
-                Timber.tag(TAG).e(e)
+                Timber.e(e)
                 Resource.error(R.string.update_item_failure_generic)
             }
         }
@@ -178,25 +168,24 @@ class ItemRepositoryImpl @Inject constructor(
     override suspend fun deleteItem(item: Item): DeleteItem {
         return withContext(ioDispatcher) {
             try {
-                // Loading state before starting the deletion operation
-                Resource.Loading
-
                 suspendCoroutine { continuation ->
+                    if (currentUserId.isBlank()) {
+                        continuation.resume(Resource.error(com.casecode.pos.domain.R.string.uid_empty))
+                        return@suspendCoroutine
+                    }
                     getItemDocumentRef(uid = currentUserId, sku = item.sku).delete()
                         .addOnSuccessListener {
                             continuation.resume(Resource.success(R.string.item_deleted_successfully))
                         }.addOnFailureListener { failure ->
-                            Timber.tag(TAG).e(failure)
-
-                            val errorMessage = when (failure) {
-                                is UnknownHostException -> R.string.delete_item_failure_network
-                                else -> R.string.delete_item_failure_generic
-                            }
-                            continuation.resume(Resource.error(errorMessage))
+                            Timber.e(failure)
+                            continuation.resume(Resource.error(R.string.delete_item_failure_generic))
                         }
                 }
+
+            } catch (e: UnknownHostException) {
+                Resource.error(R.string.delete_item_failure_network)
             } catch (e: Exception) {
-                Timber.tag(TAG).e(e)
+                Timber.e(e)
                 Resource.error(R.string.delete_item_failure_generic)
             }
         }
@@ -213,10 +202,8 @@ class ItemRepositoryImpl @Inject constructor(
         uid: String,
         sku: String,
     ): DocumentReference {
-        val itemsCollection =
-            firestore.collection(USERS_COLLECTION_PATH)
-                .document(uid)
-                .collection(ITEMS_COLLECTION_PATH)
+        val itemsCollection = firestore.collection(USERS_COLLECTION_PATH).document(uid)
+            .collection(ITEMS_COLLECTION_PATH)
         return itemsCollection.document(sku)
     }
 
@@ -227,10 +214,11 @@ class ItemRepositoryImpl @Inject constructor(
      * @return The Firestore collection reference.
      */
     private fun getItemCollectionRef(uid: String): CollectionReference =
-        firestore.collection(USERS_COLLECTION_PATH).document(uid)
-            .collection(ITEMS_COLLECTION_PATH)
+        firestore.collection(USERS_COLLECTION_PATH).document(uid).collection(ITEMS_COLLECTION_PATH)
 
     companion object {
-        private val TAG = ItemRepositoryImpl::class.java.simpleName
+        private const val LOCAL_CACHE = "local_cache"
+        private const val LOCAL_DB = "Local_DB"
+        private const val SERVER_DB = "Server_DB"
     }
 }
